add if condition for auto-start

+++ kubernetes/recipes/flanneld.rb
@@ -2,11 +2,13 @@ bash 'install_flannel' do
   user 'root'
   cwd '/tmp'
   code <<-EOH
-  wget --max-redirect 255 https://github.com/coreos/flannel/releases/download/v0.5.2/flannel-0.5.2-linux-amd64.tar.gz
-  tar zxvf flannel-0.5.2-linux-amd64.tar.gz
-  cd flannel-0.5.2
-  cp flanneld /usr/local/bin
-  cp mk-docker-opts.sh /opt/
+  if [ -f /usr/local/bin/flanneld ]; then
+    wget --max-redirect 255 https://github.com/coreos/flannel/releases/download/v0.5.2/flannel-0.5.2-linux-amd64.tar.gz
+    tar zxvf flannel-0.5.2-linux-amd64.tar.gz
+    cd flannel-0.5.2
+    cp flanneld /usr/local/bin
+    cp mk-docker-opts.sh /opt/
+  fi
   EOH
 end
